+++ Changelog.md
@@ -7,6 +7,7 @@
   - Chicago, Illinois (#220).
   - Guam territory (#219).
   - Suffolk County, Massachusetts (#222).
+  - Cayman Islands, British Overseas Territory (#328)
 
 ## v4.1.0 (2019-02-07)